Optimize filter and metaData.
