+++ wormhole-features/wormhole-console/src/main/resources/Mapper/UserInfoMapper.xml
@@ -23,7 +23,7 @@
         <result column="username" property="username" jdbcType="VARCHAR"/>
         <result column="password" property="password" jdbcType="VARCHAR"/>
         <result column="email" property="email" jdbcType="VARCHAR"/>
-        <result column="enable" property="enable" jdbcType="INTEGER"/>
+        <result column="is_enable" property="enable" jdbcType="INTEGER"/>
         <result column="create_time" property="createTime" jdbcType="TIMESTAMP"/>
         <result column="modify_time" property="modifyTime" jdbcType="TIMESTAMP"/>
     </resultMap>
@@ -36,27 +36,17 @@
         id = #{id,jdbcType=INTEGER}
     </sql>
 
-    <sql id="where_enable">
-        <if test="null != status">
-            AND enable = #{enable,jdbcType=INTEGER}
-        </if>
-    </sql>
-
     <sql id="limit">
         LIMIT #{start}, #{size}
     </sql>
 
     <sql id="insert_columns">
-        username, password, email, enable, create_time, modify_time
+        username, password, email, is_enable, create_time, modify_time
     </sql>
 
     <sql id="insert_values">
         #{username,jdbcType=VARCHAR}, #{password,jdbcType=VARCHAR}, #{email,jdbcType=VARCHAR},
         #{enable,jdbcType=INTEGER}, #{createTime,jdbcType=TIMESTAMP}, #{modifyTime,jdbcType=TIMESTAMP}
-    </sql>
-
-    <sql id="select_columns">
-        id, username, password, email, enable, create_time, modify_time
     </sql>
 
     <sql id="update_columns">
@@ -70,7 +60,21 @@
             email = #{email,jdbcType=VARCHAR},
         </if>
         <if test="null != enable">
-            enable = #{enable,jdbcType=INTEGER},
+            is_enable = #{enable,jdbcType=INTEGER},
+        </if>
+    </sql>
+
+    <sql id="select_columns">
+        id, username, password, email, is_enable, create_time, modify_time
+    </sql>
+
+    <sql id="select_count">
+        COUNT(id) AS total
+    </sql>
+
+    <sql id="where_enable">
+        <if test="null != enable">
+            AND is_enable = #{enable,jdbcType=INTEGER}
         </if>
     </sql>
 
@@ -83,7 +87,7 @@
         </if>
     </sql>
 
-    <sql id="where_page_query">
+    <sql id="where_fuzzy_query_page">
         <if test="null != query.username">
             OR username LIKE CONCAT('%', #{query.username,jdbcType=VARCHAR}, '%')
         </if>
@@ -93,7 +97,7 @@
     </sql>
 
     <insert id="insert" parameterType="com.zergclan.wormhole.console.application.domain.entity.UserInfo">
-        INSERT INTO user_info (<include refid="insert_columns"/>) VALUES (<include refid="insert_values"/>)
+        INSERT INTO <include refid="table_name"/> (<include refid="insert_columns"/>) VALUES (<include refid="insert_values"/>)
     </insert>
 
     <update id="updateById" parameterType="com.zergclan.wormhole.console.application.domain.entity.UserInfo">
@@ -127,7 +131,7 @@
     </select>
 
     <select id="countByPage" parameterType="com.zergclan.wormhole.console.application.domain.entity.UserInfo" resultType="java.lang.Integer">
-        SELECT COUNT(id) AS total FROM <include refid="table_name"/>
+        SELECT <include refid="select_count"/> FROM <include refid="table_name"/>
         <where>
             <include refid="where_fuzzy_query"/>
         </where>
@@ -136,7 +140,7 @@
     <select id="page" parameterType="com.zergclan.wormhole.console.api.vo.PageQuery" resultMap="UserInfoResultMap">
         SELECT <include refid="select_columns"/> FROM <include refid="table_name"/>
         <where>
-            <include refid="where_page_query"/>
+            <include refid="where_fuzzy_query_page"/>
         </where>
         <include refid="limit"/>
     </select>
