--- wormhole-infra/wormhole-core/src/main/java/com/zergclan/wormhole/core/metadata/ColumnMetaData.java
@@ -1,45 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.zergclan.wormhole.core.metadata;
-
-import com.zergclan.wormhole.common.SystemConstant;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-/**
- * Meta data for column.
- */
-@RequiredArgsConstructor
-@Getter
-public final class ColumnMetaData implements MetaData {
-
-    private final String owner;
-
-    private final String name;
-
-    private final int dataType;
-
-    private final boolean primaryKey;
-
-    private final boolean uniqueIndex;
-    
-    @Override
-    public String getIdentifier() {
-        return owner + SystemConstant.IDENTIFIER_SPACE + name;
-    }
-}
