--- wormhole-infra/wormhole-core/src/main/java/com/zergclan/wormhole/pipeline/converter/DataNodeConverter.java
@@ -1,37 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.zergclan.wormhole.pipeline.converter;
-
-import com.zergclan.wormhole.pipeline.DataNodeFilter;
-
-/**
- * The root interface from which all data node converter objects shall be derived in Wormhole.
- *
- * @param <S> class type of source data node
- * @param <T> class type of target data node
- */
-public interface DataNodeConverter<S, T> extends DataNodeFilter {
-    
-    /**
-     * Convert data node.
-     *
-     * @param dataNode source data node
-     * @return target data node
-     */
-    T convert(S dataNode);
-}
