+++ wormhole-infra/wormhole-core/src/main/java/com/zergclan/wormhole/pipeline/filter/StringToIntegerConverter.java
@@ -1,0 +1,46 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.zergclan.wormhole.pipeline.filter;
+
+import com.zergclan.wormhole.common.WormholeException;
+import com.zergclan.wormhole.core.metadata.DataNode;
+import com.zergclan.wormhole.core.metadata.data.IntegerDataNode;
+import com.zergclan.wormhole.core.metadata.data.StringDataNode;
+import com.zergclan.wormhole.pipeline.DataNodeFilter;
+
+/**
+ * String to integer converter for data node.
+ */
+public final class StringToIntegerConverter implements DataNodeFilter<DataNode<?>> {
+    
+    @Override
+    public DataNode<?> doFilter(final DataNode<?> node) {
+        if (node instanceof StringDataNode) {
+            StringDataNode stringDataNode = (StringDataNode) node;
+            String stringValue = stringDataNode.getValue();
+            try {
+                Integer integerValue = Integer.parseInt(stringValue);
+                IntegerDataNode integerDataNode = new IntegerDataNode(stringDataNode.getName());
+                return integerDataNode.refresh(integerValue);
+            } catch (NumberFormatException ex) {
+                throw new WormholeException(ex.getCause());
+            }
+        }
+        throw new WormholeException("StringToIntegerConverter handler exception");
+    }
+}
